+++ tidy3d/plugins/expressions/metrics.py
@@ -1,11 +1,10 @@
-from typing import Any
+from typing import Any, Optional, Union
 
-import autograd.numpy as anp
+import autograd.numpy as np
 import pydantic.v1 as pd
 
 from tidy3d.components.monitor import ModeMonitor
 from tidy3d.components.types import Direction, FreqArray
-from tidy3d.exceptions import ValidationError
 
 from .types import NumberType
 from .variables import Variable
@@ -27,48 +26,51 @@ class ModeAmp(Metric):
     """
     Metric for calculating the mode coefficient from a ModeMonitor.
 
-    Parameters
-    ----------
-    monitor_name : str
-        The name of the mode monitor.
-    freqs : FreqArray
-        The frequency array.
-    direction : Direction = "+"
-        The direction of the mode.
-    mode_index : pd.NonNegativeInt = 0
-        The index of the mode.
-
     Examples
     --------
-    >>> monitor = ModeMonitor(name="monitor1", freqs=[1.0])
+    >>> monitor = ModeMonitor(name="monitor1", f=[1.0])
     >>> mode_coeff = ModeAmp.from_mode_monitor(monitor)
     >>> data = SimulationData()  # Assume this is a valid SimulationData object
     >>> result = mode_coeff.evaluate(data)
     """
 
-    monitor_name: str
-    freqs: FreqArray
-    direction: Direction = "+"
-    mode_index: pd.NonNegativeInt = 0
-
-    @pd.validator("freqs", always=True)
-    def _single_frequency(cls, val: FreqArray) -> FreqArray:
-        if len(val) != 1:
-            raise ValidationError("Only a single frequency is supported at the moment.")
-        return val
+    monitor_name: str = pd.Field(
+        ...,
+        title="Monitor Name",
+        description="The name of the mode monitor. This needs to match the name of the monitor in the simulation.",
+    )
+    f: Optional[Union[float, FreqArray]] = pd.Field(  # type: ignore
+        None,
+        title="Frequency Array",
+        description="The frequency array. If None, all frequencies in the monitor will be used.",
+    )
+    direction: Direction = pd.Field(
+        "+",
+        title="Direction",
+        description="The direction of propagation of the mode.",
+    )
+    mode_index: pd.NonNegativeInt = pd.Field(
+        0,
+        title="Mode Index",
+        description="The index of the mode.",
+    )
 
     @classmethod
-    def from_mode_monitor(cls, monitor: ModeMonitor):
-        return cls(monitor_name=monitor.name, freqs=monitor.freqs, mode_index=0)
+    def from_mode_monitor(
+        cls, monitor: ModeMonitor, mode_index: int = 0, direction: Direction = "+"
+    ):
+        return cls(
+            monitor_name=monitor.name, f=monitor.freqs, mode_index=mode_index, direction=direction
+        )
 
     def evaluate(self, *args: Any, **kwargs: Any) -> NumberType:
         data = super().evaluate(*args, **kwargs)
-        amps = (
-            data[self.monitor_name]
-            .amps.sel(direction=self.direction, mode_index=self.mode_index)
-            .isel(f=0)
+        amps = data[self.monitor_name].amps.sel(
+            direction=self.direction, mode_index=self.mode_index
         )
-        return anp.squeeze(amps.values.tolist())
+        if self.f is not None:
+            amps = amps.sel(f=list(self.f), method="nearest")
+        return np.squeeze(amps.values.tolist())
 
 
 class ModePower(ModeAmp):
@@ -77,7 +79,7 @@ class ModePower(ModeAmp):
 
     Examples
     --------
-    >>> monitor = ModeMonitor(name="monitor1", freqs=[1.0])
+    >>> monitor = ModeMonitor(name="monitor1", f=[1.0])
     >>> mode_power = ModePower.from_mode_monitor(monitor)
     >>> data = SimulationData()  # Assume this is a valid SimulationData object
     >>> result = mode_power.evaluate(data)
@@ -85,4 +87,4 @@ class ModePower(ModeAmp):
 
     def evaluate(self, *args: Any, **kwargs: Any) -> NumberType:
         amps = super().evaluate(*args, **kwargs)
-        return abs(amps) ** 2
+        return np.abs(amps) ** 2